Respond to latest feedback; fix minor Rose file recognition bug
