+++ lib/cylc/cylc-review/template/suites.html
@@ -155,14 +155,6 @@ class="table table-bordered table-condensed table-striped">
       <th>
         last active time
       </th>
-
-      <th>
-        project
-      </th>
-
-      <th>
-        title
-      </th>
     </tr>
   </thead>
   <tbody>
@@ -187,10 +179,6 @@ class="table table-bordered table-condensed table-striped">
           title="{{entry.last_activity_time}}">{{entry.last_activity_time}}</abbr>
         </small>
       </td>
-
-      <td><small>{{entry.info["project"]}}</small></td>
-
-      <td><small>{{entry.info["title"]}}</small></td>
 
     </tr>
 {% endfor -%}{# entry in entries #}
